Update Capstone1_Data_Story_Code.ipynb
BC it's not in a pull request. 
But, Lauren
your variety in graphs and complete/clear comments reflect the traits of an individual considered valuable in the data science/analytics field. 

"We can see above that the median price of a single family home appears to be close to the median price for a townhouse. We may explore this further in the future." <---- this is interesting